advances

+++ src/main/java/com/naturalia/backend/service/impl/CategoryServiceImpl.java
@@ -2,6 +2,7 @@ package com.naturalia.backend.service.impl;
 
 import com.naturalia.backend.entity.Category;
 import com.naturalia.backend.repository.ICategoryRepository;
+import com.naturalia.backend.repository.IStayRepository;
 import com.naturalia.backend.service.ICategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,7 @@ import java.util.List;
 public class CategoryServiceImpl implements ICategoryService {
 
     private final ICategoryRepository categoryRepository;
+    private final IStayRepository iStayRepository;
 
     @Override
     public Category createCategory(Category category) {
@@ -24,8 +26,18 @@ public class CategoryServiceImpl implements ICategoryService {
         return categoryRepository.findAll();
     }
 
+
+
     @Override
     public void deleteCategory(Long id) {
+        Category category = categoryRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Categoría no encontrada"));
+
+        boolean isUsed = iStayRepository.existsByCategory(category);
+        if (isUsed) {
+            throw new RuntimeException("No se puede eliminar: categoría asociada a uno o más productos");
+        }
+
         categoryRepository.deleteById(id);
     }
 }
